+++ R/data.R
@@ -1,6 +1,10 @@
 
 
-#' Dummy text 1
+#' Demo data set UIBK (XLSX)
+#'
+#' One of the demo data sets used for testing and for the documentation/manuals.
+#' An XLSX file containing both, the measurements as well as a configuration to
+#' prepare the annex objects.
 #'
 #' @name demo_UIBK
 #' @docType data
@@ -8,7 +12,13 @@
 #' @keywords data
 NULL
 
-#' Dummy text 2
+#' Demo data set UIBK config (textfile)
+#'
+#' One of the demo data sets used for testing and for the documentation/manuals.
+#' A tabular text file (CSV alike) containing the config information
+#' for the measurement sheet in the \code{demo_UIBK} data set.
+#' An alternative to the configuration contained in a dedicated sheet
+#' in the \code{demo_UIBK} data set.
 #'
 #' @name demo_UIBK_config
 #' @docType data
@@ -16,7 +26,12 @@ NULL
 #' @keywords data
 NULL
 
-#' Dummy text 3
+#' Demo data set Bedroom
+#'
+#' One of the demo data sets used for testing and for the documentation/manuals.
+#' At tabular text file (CSV alike) containing measurement data.
+#' The data set \code{demo_Bedroom_config} contains the corresponding
+#' configuration used when calling [annex::annex_prepare()].
 #'
 #' @name demo_Bedroom
 #' @docType data
@@ -25,6 +40,10 @@ NULL
 NULL
 
 #' Dummy text 4
+#'
+#' One of the demo data sets used for testing and for the documentation/manuals.
+#' A tabular text file (CSV alike) containing the config for the
+#' \code{demo_Bedroom} data set, which contains the actual measurements.
 #'
 #' @name demo_Bedroom_config
 #' @docType data