+++ R/globals.R
@@ -1,4 +1,5 @@
 
 
 # Global bindings for package checks
-utils::globalVariables(c("variable", "column", "study", "home", "room", "config"))
+utils::globalVariables(c("variable", "column", "study", "home", "room", "config",
+                         "ID"))